refine global buffer get

+++ lib/Dialect/Tpu/Interfaces/BM1684X/LRN.cpp
@@ -22,8 +22,8 @@ extern "C" {
 #endif
 
 typedef struct {
-  unsigned long long input_addr;
-  unsigned long long output_addr;
+  uint64_t input_addr;
+  uint64_t output_addr;
   int input_n;
   int input_c;
   int input_h;
